Clean-up wavelet decomposition and reconstruction
- gets rid of many copy operations
- makes code more readable and clear
- wavelet reconstruction code was shortened by combining the cases for even or odd g
- one definition change: Spaghetti format changed from HH HG GH GG to HH GH HG GG

Explanation on definition change:
the first fast-changing index reflects the x dimension, it would make no sense if in this dimension we apply the H or G filter in the y-dimension. While before the naming was more convenient with SC, WCx, WCy, WCxy; I think this definition is not consistent with indexing. This was handled within the code by a mere swapping of indices after (or before for reconstruciton) the filter operations were applied.